Fake project about money converter from USD > IDR
heyho guys, so this time I have studied and made a series of coding (but still a fake project) about converting from US dollars to Indonesian Rupiah. In this coding there are lots of things that I already understand and don't understand either, for example those who already understand, namely from making the GUI with tkinter, then how to get it out and there's still a lot. Thank You